app/src/requirements.txt
Flask==2.2.2
flask-restx==1.0.3
Flask-JWT-Extended==4.4.4
gevent==22.10.2
python-dotenv==0.21.0
flask-sqlalchemy==3.0.2
Flask-Migrate==4.0.0
psycopg2-binary==2.9.5
passlib==1.7.4
argon2-cffi==21.3.0
redis==4.4.0
click==8.1.3
Authlib==1.2.0


